fix 🐛: unclustered root.go
