+++ cmd/root.go
@@ -1,14 +1,11 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
 
 	"log"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"strings"
 
 	"github.com/muandane/goji/pkg/config"
 	"github.com/muandane/goji/pkg/utils"
@@ -37,28 +34,20 @@ var rootCmd = &cobra.Command{
 			color.Unset()
 			return
 		}
+		if initFlag {
+			color.Set(color.FgYellow)
+			err := config.InitRepoConfig()
+			if err != nil {
+				fmt.Printf("Failed to initialize repository configuration: %v\n", err)
+				return
+			}
+			color.Unset()
+			return
+		}
+
 		color.Set(color.FgGreen)
 		fmt.Printf("Goji v%s is a cli tool to generate conventional commits with emojis.\n", version)
 		color.Unset()
-
-		if initFlag {
-			gitmojis := AddCustomCommitTypes([]Gitmoji{})
-			config := initConfig{
-				Types:            gitmojis,
-				Scopes:           []string{"home", "accounts", "ci"},
-				Symbol:           true,
-				SkipQuestions:    []string{},
-				SubjectMaxLength: 50,
-			}
-			err := SaveGitmojisToFile(config, ".goji.json")
-			if err != nil {
-				fmt.Printf("Error saving gitmojis to file: %v\n", err)
-				os.Exit(1)
-			}
-
-			fmt.Println("Gitmojis saved to .goji.json 🎊")
-			return
-		}
 
 		if typeFlag != "" && messageFlag != "" {
 			config, err := config.LoadConfig(".goji.json")
@@ -134,64 +123,6 @@ func init() {
 	rootCmd.Flags().StringVarP(&messageFlag, "message", "m", "", "Specify a commit message")
 }
 
-type Gitmoji struct {
-	Emoji       string `json:"emoji"`
-	Code        string `json:"code"`
-	Description string `json:"description"`
-	Name        string `json:"name"`
-}
-type initConfig struct {
-	Types            []Gitmoji `json:"Types"`
-	Scopes           []string  `json:"Scopes"`
-	Symbol           bool      `json:"Symbol"`
-	SkipQuestions    []string  `json:"SkipQuestions"`
-	SubjectMaxLength int       `json:"SubjectMaxLength"`
-}
-
-func AddCustomCommitTypes(gitmojis []Gitmoji) []Gitmoji {
-	customGitmojis := []Gitmoji{
-		{Emoji: "✨", Code: ":sparkles:", Description: "Introduce new features.", Name: "feat"},
-		{Emoji: "🐛", Code: ":bug:", Description: "Fix a bug.", Name: "fix"},
-		{Emoji: "📚", Code: ":books:", Description: "Documentation change.", Name: "docs"},
-		{Emoji: "🎨", Code: ":art:", Description: "Improve structure/format of the code.", Name: "refactor"},
-		{Emoji: "🧹", Code: ":broom:", Description: "A chore change.", Name: "chore"},
-		{Emoji: "🧪", Code: ":test_tube:", Description: "Add a test.", Name: "test"},
-		{Emoji: "🚑️", Code: ":ambulance:", Description: "Critical hotfix.", Name: "hotfix"},
-		{Emoji: "⚰️", Code: ":coffin:", Description: "Remove dead code.", Name: "deprecate"},
-		{Emoji: "⚡️", Code: ":zap:", Description: "Improve performance.", Name: "perf"},
-		{Emoji: "🚧", Code: ":construction:", Description: "Work in progress.", Name: "wip"},
-		{Emoji: "📦", Code: ":package:", Description: "Add or update compiled files or packages.", Name: "package"},
-	}
-
-	return append(gitmojis, customGitmojis...)
-}
-
-func GetGitRootDir() (string, error) {
-	gitRoot := exec.Command("git", "rev-parse", "--show-toplevel")
-	gitDirBytes, err := gitRoot.Output()
-	if err != nil {
-		return "", fmt.Errorf("error finding git root directory: %v", err)
-	}
-	gitDir := string(gitDirBytes)
-	gitDir = strings.TrimSpace(gitDir) // Remove newline character at the end
-
-	return gitDir, nil
-}
-
-func SaveGitmojisToFile(config initConfig, filename string) error {
-	gitDir, err := GetGitRootDir()
-	if err != nil {
-		return err
-	}
-
-	configFile := filepath.Join(gitDir, filename)
-	data, err := json.MarshalIndent(config, "", "  ")
-	if err != nil {
-		return err
-	}
-
-	return os.WriteFile(configFile, data, 0644)
-}
 func Execute() {
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
