+++ pkg/config/configType.go
@@ -10,3 +10,17 @@ type Config struct {
 	Questions        map[string]string
 	SubjectMaxLength int
 }
+
+type Gitmoji struct {
+	Emoji       string `json:"emoji"`
+	Code        string `json:"code"`
+	Description string `json:"description"`
+	Name        string `json:"name"`
+}
+type initConfig struct {
+	Types            []Gitmoji `json:"Types"`
+	Scopes           []string  `json:"Scopes"`
+	Symbol           bool      `json:"Symbol"`
+	SkipQuestions    []string  `json:"SkipQuestions"`
+	SubjectMaxLength int       `json:"SubjectMaxLength"`
+}
